trend_research_pipeline/requirements.txt
browserbase>=0.3.0
selenium>=4.15.0
python-dotenv>=1.0.0
schedule>=1.2.0
